+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [1.3.1] - 2023-06-10
+* Add `Navigatable::enterNavigated()`, `Navigatable::resetNavigation()`,
+  `Navigatable::isNavigated()`
+* Remove unused `Navigatable::autoDeactivation_`
+
 ## [1.3.0] - 2023-06-10
 * Add class `BWidgets::Activatable` and implement in `BWidgets::Widget`
 * Add class `BWidgets::Navigatable` and implement in `BWidgets::Box`